// src/app/views/models/registroAba.model.ts
export interface Paciente {
    value: string;
    viewValue: string;
}

export interface Programa {
    value: string;
    viewValue: string;
}

export interface Legendas {
    value: string;
    viewValue: string;
}

export interface Paciente {
    value: string;
    viewValue: string;
  }
  
  export interface Programas{
    value: string;
    viewValue: string;
  }
  
  export interface Programas_paciente {
    position: number;
    programa: string;
    numero_de_tentativas: number;
  }